Better error handling
Wrapping errors when it makes sense (near the root), to add additional
context to the user not running with -vv.
(For instance "couldn't get where to save reported metrics on disk" +
"coudn't get home dir…" instead of only "couldn't get home dir…").

Start protecting errorFormat to be spread out in the code.