generate commands in README.md from cobra output
That enables us to stay in sync with the command syntax by simply running
go test --generate in the cmd/ubuntu-report package.
We rewrite on demand the README.md file, skipping the previous usage section
and do this on all subcommands, only extracting the relevant part (skipping
SEE ALSO and such). Finally, we skip the interactive details as it's an alias
to running "ubuntu-report".