Ensure we can mock Metrics elements outside of the package
Unfortunately, we have to ship the Mock to the external API (and not only
our modifiers in _test.go), as we want to mock the whole system in run.go
from another package.
As told in previous commit, this enables us parallelized testing of run
functions, isolated from base environment.