tests for C API
Adapt a similar technic than in Go Stdlib to use C API from withing go
and test it under the same condition.
As the C API isn't imported anywhere, the cgo import won't be in pure
GO binary.
I won't be even in the shared lib as we specify the file, and we did
our tests and additional export definition in another file.