Add vendoring in debian package
We thought using the distribution version for dep was the best path
for ubuntu and favored it over using our vendor deps.
It seems that vendor deps are leaking in built-using though, and so all
build-deps would need to be promoted in main.

As this seems undesired, move back to using vendor deps (which are thus a copy
of the exact same code, in different versions) as this is what others projects
are already using.