Report disks size
Compute it from the logical block size and number of blocks for each
local device.
A local device is defined with name matching {h;s;v}d*, which should
handle traditional spinning disks and old installations, ssd and vm
virtual disks.
The report is in Gib, as for partitions size.