autostart upgrade mode trigger
Try to start ubuntu-report in dist-upgrade mode. As the report file as a
per-release file name, we can't depend on it in an AutostartCondition or
it will mean having to rebuild the package once per release.
We prefer thus to execute it unconditionnally and exit quickly… lacking
for a better option.